# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/application/code": {
            "post": {
                "description": "用户点击\"获取验证码\"按钮，系统向用户提供的邮箱发送6位验证码，用户需要在申请表单中填入验证码才可以成功完成身份验证，否则不应该可以提交申请。验证码时限为10分钟，超时无效",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理"
                ],
                "summary": "获取申请验证码 Vera",
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/response.GetVerifyCodeQ"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\": \"邮件发送成功\",\"status\": 200}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"msg\": \"数据格式错误\", \"status\": 400}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "{\"msg\": \"没有该用户\", \"status\": 401}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "402": {
                        "description": "{\"msg\": \"验证码存储失败\",\"status\": 402}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "{\"msg\": \"发送邮件失败\",\"status\": 403}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/application/create": {
            "post": {
                "description": "用户可以申请认领自己的学者门户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理"
                ],
                "summary": "申请学者门户 Vera \u0026 txc",
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/response.CreateApplicationQ"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\": \"申请成功\", \"status\": 200,\"application\": application}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"msg\": \"数据格式错误\", \"status\": 400}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "{\"msg\": \"没有该用户\", \"status\": 401}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "402": {
                        "description": "{\"msg\": \"创建申请失败\", \"status\": 402}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "{\"msg\": \"该学者已被认领\", \"status\": 403}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/application/handle": {
            "post": {
                "description": "管理员对用户提交的申请进行审核，并给出审核意见content",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理"
                ],
                "summary": "审核学者门户申请 Vera \u0026 txc",
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/response.HandleApplicationQ"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\": \"审核成功\", \"status\": 200}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"msg\": \"数据格式错误\", \"status\": 400}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "{\"msg\": \"管理员不存在 or 不是管理员\", \"status\": 401}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "402": {
                        "description": "{\"msg\": \"申请不存在\", \"status\": 402}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "{\"msg\": \"已审核过该申请\", \"status\": 403}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "{\"msg\": \"数据库错误\", \"status\": 404}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/es/get/": {
            "get": {
                "description": "根据id获取对象，可以是author，work，institution,venue,concept W4237558494,W2009180309,W2984203759",
                "tags": [
                    "esSearch"
                ],
                "summary": "根据id获取对象 txc",
                "parameters": [
                    {
                        "type": "string",
                        "description": "对象id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户id",
                        "name": "userid",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\":200,\"res\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"status\":400,\"msg\":\"id type error\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "{\"status\":201,\"msg\":\"es get err or not found\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/es/get2/": {
            "get": {
                "description": "根据id获取对象，可以是author，work，institution,venue,concept W4237558494,W2009180309,W2984203759",
                "tags": [
                    "esSearch"
                ],
                "summary": "根据id获取对象 txc",
                "parameters": [
                    {
                        "type": "string",
                        "description": "对象id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户id",
                        "name": "userid",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\":200,\"res\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"status\":400,\"msg\":\"id type error\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "{\"status\":201,\"msg\":\"es get err or not found\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/es/getAuthorRelationNet": {
            "get": {
                "description": "根据author的id获取专家关系网络, 目前会返回Top N的关系网，N=10，后续可以讨论修改N的大小或者传参给我\n\n目前接口时延约为1s, 后续考虑把计算出来的结果存入数据库，二次查询时延降低\n",
                "tags": [
                    "esSearch"
                ],
                "summary": "根据author的id获取专家关系网络 hr",
                "parameters": [
                    {
                        "enum": [
                            "A2764814280",
                            "A2900471938",
                            "A2227665069"
                        ],
                        "type": "string",
                        "description": "author_id",
                        "name": "author_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"res\":{ \"Vertex_set\":[], \"Edge_set\":[]}}",
                        "schema": {
                            "$ref": "#/definitions/response.AuthorRelationNet"
                        }
                    },
                    "201": {
                        "description": "{\"msg\":\"Get Author Relation Net Error\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/es/prefix": {
            "post": {
                "description": "根据前缀得到搜索建议，返回results 字符串数组",
                "tags": [
                    "esSearch"
                ],
                "summary": "根据前缀得到搜索建议，返回results 字符串数组 hr",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Field 表示需要查询的字段名",
                        "name": "Field",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Prefix 表示用户已经输入的前缀",
                        "name": "Prefix",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\": true, \"msg\": \"获取成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"success\": false, \"msg\": 参数错误\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "402": {
                        "description": "{\"success\": false, \"msg\": \"es服务出错\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/es/search/advanced": {
            "post": {
                "description": "高级搜索，query是一个map列表， 每个map包含\"content\" \"field\" \"logic\"\nlogic 仅包含[\"and\", \"or\", \"not\"]\nfield 仅包含[\"title\", \"abstract\", \"venue\", \"publisher\", \"author\", \"institution\", \"concept\"]\n对于年份的筛选，在query里面 field是\"publication_date\" logic默认为and， 该map下有\"begin\" \"end\"分别是开始和结束\nsort=0为默认排序（降序） =1为按引用数降序 =2按发表日期由近到远\nasc=0为降序 =1为升序\n{ \"asc\": false,\"conds\": {\"venue\":\"International Journal for Research in Applied Science and Engineering Technology\",\"author\": \"Zenith Nandy\"},\"page\": 1,\"query\": [{\"field\": \"title\",\"content\": \"python\",\"logic\": \"and\"},{\"field\": \"publication_date\",\"begin\": \"2021-12-01\",\"end\":\"2022-06-01\",\"logic\": \"and\"}],\"size\": 8,\"sort\": 0}",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "esSearch"
                ],
                "summary": "高级搜索 txc",
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/response.AdvancedSearchQ"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/es/search/author": {
            "get": {
                "description": "根据作者名字搜索作者,模糊搜索",
                "tags": [
                    "esSearch"
                ],
                "summary": "txc",
                "parameters": [
                    {
                        "type": "string",
                        "description": "query_word",
                        "name": "query_word",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "size",
                        "name": "size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "sort",
                        "name": "sort",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "asc",
                        "name": "asc",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"res\":{},\"msg\": \"Author Search Success\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "{\"msg\": \"Author Search Error\",\"err\":err}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/es/search/author2": {
            "get": {
                "description": "根据作者名字搜索作者 via openalex",
                "tags": [
                    "esSearch"
                ],
                "summary": "txc",
                "parameters": [
                    {
                        "type": "string",
                        "description": "query_word",
                        "name": "query_word",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "size",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "sort=cited_by_count|...:_|desc|asc",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"res\":{},\"msg\": \"Author Search Success\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "{\"msg\": \"openalex Search Error\",\"err\":err}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "402": {
                        "description": "{\"msg\": \"openalex Search Error\",\"err\":err}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/es/search/base": {
            "post": {
                "description": "基本搜索，Cond里面填筛选条件，key仅包含[\"type\", \"author\", \"institution\", \"publisher\", \"venue\", \"publication_year\"]",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "esSearch"
                ],
                "summary": "基本搜索 txc",
                "parameters": [
                    {
                        "description": "搜索条件",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/response.BaseSearchQ"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\":200,\"res\":{obeject}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "201": {
                        "description": "{\"status\":201,\"err\":\"es search err\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/es/search/doi": {
            "post": {
                "description": "使用doi查找work，未测试，请勿使用",
                "tags": [
                    "esSearch"
                ],
                "summary": "使用doi查找work，未测试，请勿使用 txc",
                "parameters": [
                    {
                        "type": "string",
                        "description": "doi",
                        "name": "doi",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/es/statistic": {
            "get": {
                "description": "获取统计信息",
                "tags": [
                    "esSearch"
                ],
                "summary": "获取统计信息 txc",
                "responses": {
                    "200": {
                        "description": "{\"res\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "301": {
                        "description": "{\"err\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/login": {
            "post": {
                "description": "登录\n填入用户名和密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "登录 ccf",
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/response.LoginQ"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\":200,\"msg\":\"登录成功\",\"token\": token,\"ID\": user.UserID}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"status\":400,\"msg\":\"用户名不存在\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "{\"status\":401,\"msg\":\"密码错误\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "注册\n填入用户名和密码注册",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "注册 ccf",
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/response.RegisterQ"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"status\":200,\"msg\":\"注册成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"status\":400,\"msg\":\"用户名已存在\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/scholar/author/headshot": {
            "post": {
                "description": "上传作者头像",
                "tags": [
                    "scholar"
                ],
                "summary": "上传作者头像 txc",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "学者ID",
                        "name": "author_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "新头像",
                        "name": "Headshot",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"上传成功\",\"data\": author}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"msg\":\"学者未被认领\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "{\"msg\":\"无权限\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "402": {
                        "description": "{\"msg\":\"头像文件获取失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/scholar/author/intro": {
            "post": {
                "description": "修改作者简介",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scholar"
                ],
                "summary": "txc",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/response.ModifyAuthorIntroQ"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"修改成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"msg\":\"参数错误\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "{\"msg\":\"无权限\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "{\"msg\":\"学者未被认领\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/scholar/hot": {
            "get": {
                "description": "获取热门论文（根据访问量）",
                "tags": [
                    "scholar"
                ],
                "summary": "获取热门论文（根据访问量） txc",
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"获取成功\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"msg\":\"获取失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.CreateApplicationQ": {
            "type": "object",
            "required": [
                "author_id",
                "author_name",
                "email",
                "institution",
                "user_id",
                "verify_code"
            ],
            "properties": {
                "author_id": {
                    "type": "string"
                },
                "author_name": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "institution": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "verify_code": {
                    "type": "string"
                }
            }
        },
...
        "response.GetVerifyCodeQ": {
            "type": "object",
            "required": [
                "email",
                "user_id"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "response.HandleApplicationQ": {
            "type": "object",
            "required": [
                "application_id",
                "status",
                "user_id"
            ],
            "properties": {
                "application_id": {
                    "type": "integer"
                },
                "content": {
                    "description": "审批意见",
                    "type": "string"
                },
                "status": {
                    "description": "是否通过",
                    "type": "integer"
                },
                "user_id": {
                    "description": "谁审核的",
                    "type": "integer"
                }
            }
        },
...
    }
}